backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt
Pillow==9.4.0
django-multiselectfield==0.1.12
drf-spectacular
